Tentative LPC1788 support
Note! There are a lot of assumptions and bad coding ahoy, review
thoroughly and expect not to be complete!

This version has most of the core files updates to support the LPC1788
and appears to compile with both mbed python scripts and Keil uVision
set to use the ARM-EABI-NONE- toolchain (although only with a lot of
wrestling), however it fails the basic test with Keil (Hard Fault when
calling wait_usec), I suspect this is an issue with the JTAG interface
as it only simulates Timers 0 and 1, whereas wait uses Timer 3. The
binary compiled with the mbed python scripts just appears to hang doing
nothing.

Hopefully Emilio and the team can make more sense of this.

- Jason