Refactored KLXX code
See description. Everything HAL wise went into KLXX, with target
specific in subfolders. Only interrupt and spi are in target specific
since they are target specific.

Also you cannot set the uart anymore for 9 or 10 databits (the code
supported it to set it like that, but the actually read/write only did
8-bits)