[DISCO_F407VG]: Change default FPU option to "softfp" from "hard".
This commit changes the defalt compile option in relation to FPU from "hard" to "softfp" in
Makefile template, because it would be better to maintain the consistency with "build.py" and "make.py"
that have the default option, "softfp".
If the "hard" option is needed, execute "make HARDFP=1" on command line.