Added "M4F"
Several listed as Cortex M4 cores are actually M4F cores (having a floating point unit), I did some homework to figure out which ones have FPU's. It appears that the K20D50M is the only board with a Cortex-M4 MCU that is does not have a floating point unit.