K22F support added
And that is a bit of an overstatement. But mbed blinky works :D. And
nothing else works atm. Still more than before today.

Every option in device.h is disabled (and won't work), but now they can
slowly be enabled (they will need pinouts and most some will need some
changes in the mbed api files).

One of the major issues was that the KSDK is intended to only compile
relevant peripherals, but the mbed build system does not support this.
So now for the affected peripherals (with the K22F and K64F) instead of
throwing an error when no valid CPU is found for them, it sets a define,
and the relevant .h and .c files have an #ifndef for that define. This
way it doesn't need to be manually set for each mbed target.