[NUCLEO_F4x1RE] test MBED_A7 InterruptIn failed
The test failed because it was not possible to switch off an already
active edge detection.
The reason was the GPIO HAL module driver function HAL_GPIO_Init. The
function changes only interrupt or event flags when the EXTI_MODE flag
in GPIO_Init->Mode is set. gpio_irq_set changed the event mode to
EDGE_NONE in gpio_irq_t structure but that doesn't changed the interrupt
and event flags in the CMSIS function because of the missing EXTI_MODE.
Now gpio_irq_set is able to switch off a specific or both edge
detections.
The MBED_A7 test was extended by another test to check if switch off
edge detection is OK.