[DISCO_F429ZZI] pin function definition
Added mostly all pin function (ADC, I2C, SPI, UART, PWM).

The basis of the target was DISCO_F407. The 429 has some additional pin
functions which are added with this commit. Also ports, pins and pin
functions of the larger packages are added not only the available of the
F429ZI package.