Use CLKIN_CRYSTAL as input source for PLL1
Previously the input source was set to CLKIN_MAINPLL which according to
the documentation is the output from PLL1 itself.  Before I made this
change I would often encounter HardFaults in SystemSetupClock() when
running unoptimized binaries (-O0).

I would like to have @micromint review and sign-off before this change
is merged.  It's possible this code was written this way for a reason
and it fixing my hangs is just a fluke.