Move exit override to retarget.cpp
I moved the exit handler from exit.c to retarget.cpp where the rest of
the standard C library retargetting routines are found.  The exported
makefiles already explicitly link in retarget.o but not exit.o
When building with the GNU toolchain, it is safest to link in the
object file directly for retargetting routines so that GNU linker has
to use these versions of the routines.

Testing:
I have only tested this change with GCC_ARM. It should really be
tested with more toolchains before getting merged.