[HAL][K20XX/KLXX] Fixed deepsleep power consumption when AnalogIn is used
The power consumption was reported by Paul Staron to be 100uA higher
when an AnalogIn was used previously.

Problem 1 is that 40uA was used by the async ADC clock, which is never
actually used, so it is disabled. Problem 2 is that setting it for high
speed mode increased it by another 60uA while in deepsleep. This
currently seems to me to be possibly a bug in the design, but the
workaround is checking if this is the case before going to deepsleep,
and if yes, disable it. Afterwards it is re-enabled.