workspace_tools: att HRM1017{_OTA,_BOOT} platforms
As the HRM1017 doesn't have a 32khz crystal for the LF clock, we use
an internal RC oscillator for this. mbed already supports this, however
as the bootloader uses the mbed SDK in its build, we have to have
a new bootloader hex file for the HRM1017 (and future platforms without
this crystal).

Signed-off-by: Jonathan Austin <jonathan.austin@arm.com>