Tools to automatically retrieve data from the mbed source files
This is a combined effort of mainly me (@Sissors) and @sarahmarshy. This
is really work in progress, but comitted for better cooperation options.

supported.py generates an HTML table with all (official) targets and
data from device.h. This is currently searchable and filterable (click
on column headers). Supported_html.py contains the html code used.

supported_beta.py is the next step, which does not create html (yet),
but outputs in console. It searches the ARM/uARM scatter file and from
this retrieves RAM and ROM sizes.