Nordic: add const qualifier to acceptedSpeeds in serial API
This is a small optimisation for ARM compiler toolchains: constant
arrays should be declared const explicitly, to let the compiler put them
into .rodata instead of .data. GCC automatically detects that the array
is read-only, and already puts it into flash.

This way, we free 136 bytes of RAM, which is nothing to sneeze at, given
that applications only get about 2048 bytes by default, on 16k targets.

Please note that you will need to disable data compression in armlink to
see the real difference, with the "--datacompressor off" switch, for
example.