Nordic: added the s110 scatter files
This commit adds the ability for a compile flag to be set from targets.py
which changes the soft device to s110.

The compile flag is: MCU_NORDIC_16K_S110. In order to avoid duplication
of the startup file, for the 16K parts we introduce an S110 or S130
scatter file. This means all 16K platforms must define
MCU_NORDIC_16K
and ONE of
MCU_NORDIC_16K_S110 or MCU_NORDIC_16K_S130