Move HRM and DELTA targets to use LFCLK define
After this change it will no longer be necessary to list all
the platforms that require the LFCKL to be the internal RC oscillator,
instead we specify it as a property of the platform.

Applications and libraries can use the TARGET_NRF_LFCLK_RC symbol to
filter for platforms that don't have an external crystal.

Signed-off-by: Jonathan Austin <jonathan.austin@arm.com>