Fixed linker script for LPCXpresso for EA LPC11U35 QuickStart Board
After version 7.5.0 the compiler used in LPCXpresso changed and as a
result some of the included libraries were changed. This caused a
build error when trying to compile exported mbed projects.

The error was with NewlibNano libraries (*_s.a) and as NewlibNano is
not used by mbed, this patch references the non-NewlibNano libraries
instead.

This change in the linker script has been verified on LPCXpresso versions
from 6.1.2 up to the current 7.9.0.