Remove doubling of buffer size in realiseEndpoint()
realiseEndpoint() was unnecessarily allocating twice the buffer space for each endpoint buffer.  This was presumably for the sake of the hardware SIE's double-buffering (EVEN/ODD) system, but that's vestigial - this implementation doesn't use the double-buffering capability at all, leaving the ODDRST bit in the CTL register always set.  The double-size allocation is a pure waste of memory.