Fix the use of static IP addresses
Prior to this change, trying to use a static IP address caused the
call to EthernetInterface::connect() to fail (timeout).  This was due
to a race condition between the thread that added the interface with
lwip's netif_add and the ethernet driver (in my case the k64f).  The
ethernet driver eventually tries to call the callback set for the netif
by netif_set_link_callback, but this is only set after the netif_add call.
If the netif_add_call thread hasn't been scheduled yet, the callback
does nothing, meaning we wait for the netif_linked semaphore until it
times out.

In the case of a static IP, the race condition ensured that the interface
didn't work.